[REFACTOR] Money와 관련된 클래스들을 money 패키지로 만듭니다
