+++ src/main/java/domain/IssuedLottos.java
@@ -1,5 +1,7 @@
 package domain;
 
+import domain.money.Money;
+
 import java.util.ArrayList;
 import java.util.List;
